requirements.txt
absl-py==0.14.1
backcall==0.2.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.7
chumpy==0.70
cycler==0.10.0
decorator==5.1.0
einops==0.3.2
ffmpeg-python==0.2.0
future==0.18.2
fvcore==0.1.5.post20210719
glob2==0.7
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
h5py==3.5.0
idna==3.2
imageio==2.9.0
iopath==0.1.9
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1610099769230/work
kornia==0.5.11
lpips==0.1.4
Markdown==3.3.4
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1620506037462/work
matplotlib-inline==0.1.3
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.6.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
oauthlib==3.1.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
ordered-set==4.0.2
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1610360239318/work
prompt-toolkit==3.0.21
protobuf==3.18.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch3d==0.4.0
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.15.0
scipy==1.7.1
sip==4.19.13
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1614001031686/work
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
torch==1.7.1
torchmeta==1.8.0
torchvision @ file:///tmp/build/80754af9/torchvision_1622214673755/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094706440/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1625664838498/work
traitlets==5.1.1
trimesh @ file:///home/conda/feedstock_root/build_artifacts/trimesh_1631856338394/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
yacs==0.1.6
